Move test utils inside module
So it can be used by other parties for testing purposes
